modified:   app/Http/Livewire/Front/Order/Ordersuccess.php 	modified:   app/Http/Livewire/Front/User/Userdashborad.php 	modified:   resources/views/livewire/front/user/userdashborad.blade.php

+++ app/Http/Livewire/Front/Order/Ordersuccess.php
@@ -5,6 +5,7 @@ namespace App\Http\Livewire\Front\Order;
 use App\Models\Cart;
 use Livewire\Component;
 use App\Models\DeliveryHeader;
+use App\Models\SalesHeader;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Session;
 
@@ -14,12 +15,20 @@ class Ordersuccess extends Component
     public function mount()
     {
         $status = Session::get('status');
+        $id = Session::get('id');
+        if($status == 'closeold'){
+            $this->order = SalesHeader::where('id',$id)->with('salesdetails')->first();
+        }
+        if($status == 'old'){
+            $this->order = DeliveryHeader::where('id',$id)->with('salesdetails')->first();
+        }
         if($status != true){
             redirect('/');
         }
-        $id = Session::get('id');
-        $this->order = DeliveryHeader::where('id',$id)->with('salesdetails')->first();
-        Cart::where('user_id',Auth::guard('client')->user()->id)->delete();
+        if($status == true){
+            $this->order = DeliveryHeader::where('id',$id)->with('salesdetails')->first();
+            Cart::where('user_id',Auth::guard('client')->user()->id)->delete();
+        }
     }
     public function render()
     {
